WT-3901 Fixed the log corruption bug. (#3925)
* Fixed the log corruption bug.

* Instead of updating the offset in the WT_TRACK_OP macro, update it in
the __wt_optrack_flush_buffer() function, that simplifies everything.

When closing down optrack operations, don't lose the error if we fail
to close the optrack file handle. Also, there's no need to clear the
optrack file handle explicitly (__wt_close() will clear it for us).

Add a comment on why we're not using the standard write path code when
writing an optrack buffer.