WT-3805 Avoid reading lookaside pages in fast truncate path
Allow fast-delete to skip pages that have lookaside information. The
change is to check visibility information during fast-delete on pages
with lookaside informatoin, and if there's nothing in the lookaside table
we need to instantiate in order to delete the page, allow a fast delete.
However, this requires changes to the page read path, because we can now
see deleted pages that have associated lookaside information. When that
happens, we have to instantiate the page, apply the lookaside information
and then apply the delete on top of both.