WT-3933 test/format failure illegal WT_REF.state rolling back deleted page (#3944)
* WT_3933 test/format failure illegal WT_REF.state rolling back deleted page

Back out 4fcf1f0 and 475c397, as soon as WT_REF.page_del.txnid is set
to WT_TXN_ABORTED, the page can be evicted, which potentially changes
the WT_REF state and/or discards the WT_REF.page_del structure. (Those
commits aren't necessary for correctness, they're diagnostic aids, so
they're not required.)

* Simplify __wt_delete_page_rollback(), it can return a value (although
that value will always be a panic return).

* __wt_txn_modify_page_delete() can fail after allocating a transaction
modify structure, but that could lead to an allocated page-delete modify
structure associated with a WT_REF that's not in a page-deleted state.

* Go back to marking truncates aborted.

Take care with ordering so switching the page_del transaction ID to
WT_TXN_ABORTED is the last thing rollback does to the page_del
structure.