WT-3972 Allow more than 64K cursors on a single table. (#3978)
There was a limitation introduced by the limit on the number of read locks that could be acquired at the same time. This change alters the limit a 16 bit number to a 32 bit number (~4 billion).

It's not likely to be performant to open that many cursors..

(cherry picked from commit 9fea638e63b9d066e6887b863dce9a48d1189a21)