WT-3987 Avoid reading lookaside pages in truncate fast path. (#3994)
The page's backing address can change underneath us if the parent page is
splitting. That's OK, but we can't afford to read the backing address more
than once, force the use of a local variable.

If we have a WT_REF_DELETED page that has no backing disk page, when we
are forced to instantiate it, we'll have a WT_REF.page_del structure but
it will have no WT_REF.page_del.update_list values because no updates
were ever instantiated. On rollback, check if there are updates to roll
back so we don't drop core in this case.

If a child page was part of a fast-delete, transaction rollback might
switch this page into its previous state at any time, so the delete must
be resolved.