WT-3982 Fix search_near with read-uncommitted isolation.
At low isolation levels, new records can appear in between positioning a
cursor and calling next/prev.  For this reason, search_near cannot
assume that after calling next (or prev), it will have a key larger (or
smaller) than the search key.  Loop and check in this (rare) case
instead.

This applies particularly to internal code that searches in the
lookaside table, because that access is always done at read-uncommitted
isolation.