WT-3979 Doxygen upgrade reports warnings on build.
Inserted <newline> characters in ALIAS values that immediately follow a
"@param config" statement are a problem in doxygen 1.8.14. For example,

	"configempty{2}=@param config\n"

is a problem, but

	"configempty{2}=@param config \n"

works fine.

The symptom is a subsequent line isn't read correctly, so if you have:

	* @param compressor the application-supplied compression handler
	* @configempty
	* @errors

You get:

	warning: return type of member WT_CONNECTION::add_compressor is
	not documented

I don't see a reason to insert <newline> characters into "config"
parameters when we don't insert <newline> characters into other
parameters, remove the <newline> characters causing the problem.