Lookaside sweep must not track keys across block boundaries.
When a page is evicted with history, the updates required to resolve
reads are written to the lookaside table as a "block".  When sweeping
the lookaside table, we need to be aware of block boundaries to
determine whether it is safe to discard the history.  In particular, the
presence of a "birthmark" in a block indicates that a record was written
to the backing file for the original table.

There was a bug in sweep where if a key appeared at the end of a block
(for one page) and the beginning of a block (for another page), sweep
could incorrectly remove history in the second block without a backing
record.