WT-3982 Fix transaction visibility bugs related to lookaside usage. (#3995)
There are two different issues fixed here:

* Fix search_near with read-uncommitted isolation.

At low isolation levels, new records can appear in between positioning a
cursor and calling next/prev.  For this reason, search_near cannot
assume that after calling next (or prev), it will have a key larger (or
smaller) than the search key.  Loop and check in this (rare) case
instead.

This applies particularly to internal code that searches in the
lookaside table, because that access is always done at read-uncommitted
isolation.

* Lookaside sweep must not track keys across block boundaries.

When a page is evicted with history, the updates required to resolve
reads are written to the lookaside table as a "block".  When sweeping
the lookaside table, we need to be aware of block boundaries to
determine whether it is safe to discard the history.  In particular, the
presence of a "birthmark" in a block indicates that a record was written
to the backing file for the original table.

There was a bug in sweep where if a key appeared at the end of a block
(for one page) and the beginning of a block (for another page), sweep
could incorrectly remove history in the second block without a backing
record.