WT-3987 Avoid reading lookaside pages in truncate fast path (#3996)
Fast-delete pages without backing disk images may still have lookaside
records. Skip the read, but don't skip instantiating lookaside records.

Fast-delete pages can have lookaside records, and those lookaside records
aren't limited to records that appeared in the backing disk image (if, in
fact, there was a backing disk image). Rewrite the code that instantiates
deleted pages to insert a tombstone in front of every record on the page,
not just the page entries that appeared in a disk image.