The test for a cursor moving backwards through the tree and racing with an internal page split was correct, but we were doing the test too late.
Imagine a parent page:

	| ... | a | g | i | ... |

and a child page:

	| a | b | c | d | e | f |

If we're splitting the child page into the parent, the eventual parent
page will be:

	| ... | a | c | e | g | i | ... |

and the eventual split page will be:

	| a | b |

if the cursor comes up out of the 'g' subtree (we're holding hazard
pointer on the 'g' page), decrements its page slot and moves to the 'a'
subtree, we're now holding a hazard pointer on 'a'. If we then realize the
'a' subtree is splitting, we can't recover because we've already moved
past the eventual 'c' and 'e' subtrees. Instead, we have to restart the
movement from the original coupled page, 'g'.

Instead of testing for a split after hazard-coupling to the wanted page,
do it as soon as we've retrieved the wanted page and before we release
the held page.