Instead of passing an additional "prev-race" flag to __wt_page_swap(), test the WT_READ_PREV flag. This means we'll do a more general test for internal pages splitting when moving bckwards through the tree.
Explanation: If we are returning internal pages during a walk (that is,
when doing eviction), we hazard-pointer couple to the internal parent
of the current page in order to return it. Testing WT_READ_PREV means
we'll detect and restart when the internal parent is splitting during
that ascent, instead of only when descending to an internal page.  The
change should avoid eviction skipping big blocks of records in some hot
trees, which doesn't seem like a bad thing.