WT-3959 Remove metadata entry for checkpoint timestamp if zero.
We added a "system:checkpoint" entry to the metadata to store the most
recent timestamp at which a checkpoint completed.  Current versions of
WiredTiger treat a missing entry as equivalent to a zero timestamp.

However, this metadata entry can confuse older versions of WiredTiger
(specifically, opening a backup cursor fails when the new entry exists).
Since MongoDB will always checkpoint without a timestamp as part of the
downgrade process, remove the metadata entry in that case so that older
versions of WiredTiger never see it.

Also, check for the "system:" previous generally rather than this
specific entry.  Then if future versions of WiredTiger introduce new
entries, they will be silently skipped when not expected.