WT-4019 reduce prepare and snapshot-isolation interactions in format
Only configure prepare 5% of the time. Because prepare is rarely
configured, let it override the logging and transaction timestamp
configurations (as long as they're not explicitly configured), to
ensure prepare is run enough.

Change the begin_transaction() code to not depend on #defines having
specific values.

Checking pthread locking functions for failure.

Sort structure fields in a couple of places, group thread operation
counts together.