WT-4060 Clean up __wt_getenv() error handling.
The __wt_getenv() function overloads error handling, using WT_NOTFOUND to
indicate that no environmental variable was set, when that error could be
returned from an underlying function as a real error. Use NULL/not-NULL
to communicate set/not-set for the environment variable and simplify error
handling in the callers.

Create a shared routine that gets and environment variable and checks for
any security issues, rather than repeat the code in multiple places.