WT-4069 format run exceeded 15 minutes past the maximum time
Minor rework of the __wt_page_in_func() "wait_cnt" value so it reflects
the number of times we've actually yielded, it's a useful thing to know
when debugging stuck processes.

Rename "wait_cnt" to be "yield_count" for consistency.

Rename "sleep_count" to be "sleep_usecs" -- it's not a counter, it's the
length of time we slept.