WT-4064 Relax checking of lookaside entry count. (#4077)
If there is a race between inserting and removing lookaside records (or some bug where the entry count is not updated correctly), treat underflow the same as a zero entry count (i.e., treat the lookaside table as empty).

At worst, this will leave some data in the lookaside table that could potentially be swept.  It should avoid assertion failures during tests while we search for the source of the underflow.