WT-4060 Clean up __wt_getenv() error handling. (#4073)
The __wt_getenv() function overloads error handling, using WT_NOTFOUND to
indicate that no environmental variable was set, when that error could be
returned from an underlying function as a real error. Use NULL/not-NULL
to communicate set/not-set for the environment variable and simplify error
handling in the callers.