WT-4004 format snapshot-isolation search mismatch failure
Don't perform the snapshot isolation checks after a prepare-conflict error is
returned. It works if prepare-conflict was returned for a single key/value
update, the value we read earlier in the transaction will be returned
when we check it. However, if T1 reads KeyX, then does a truncate that
includes KeyX, which successfully truncates KeyX but later fails because
of prepare-conflict, then the snapshot-isolation check fails because we see
the results of the truncate instead of the value we originally read for KeyX.