Track the boundary between on-page and lookaside history more clearly.
Skew towards old page values if checkpoints are not moving forward.
In particular, if checkpoint leaves a page dirty, track the largest
transaction it saw and use that to inform future choices about what kind
of eviction to do.  Also track whether a page was originally restored
from lookaside, and track that across update/restore eviction.