WT-4010 Simplify test/format timestamp handling.
Instead of tracking in-use timestamps, have threads update their public
timestamp information with a "last-used" timestamp, then the timestamp
thread can use the lowest "last-used" timestamp as the oldest timestamp
in the system.

Skip read-timestamp allocation if a prepare is in progress, to allow for
more parallelism between threads.

Do the final increment of the oldest timestamp as part of shutting down
the timestamp thread, don't make verify responsible for timestamp cleanup.