WT-4147 For log recovery/salvage, report/fix corruption outside of log records.
We do a conservative consistency check of log record headers, as well as
consistency checks of 'holes' in log files. This catches corruptions that
occur from appending junk to a valid log file, or inserting corruption
in the unused portion of the final log file.

Corruption errors are reported in a standard way that suggests using
salvage to fix.

Tests have been updated, along with extra commentary.