CID 104111 (#1 of 1): Redundant test (DEADCODE) dead_error_condition: The condition ret != 0 cannot be true.
Additionally, the error handling was wrong, we didn't check for errors
found while loading the data (the WT_CURSOR.next loop simply exited if
WT_NOTFOUND or some other error was returned and ignore the latter case.

Don't fix it, just convert it all to the usual test_XXX macros that
exit immediately on error. And, add an assertion that we read all of
the data, that is, that WT_CURSOR.next eventually returns WT_NOTFOUND.