WT-4187 Coverity: unused value complaints (#4179)
The WiredTiger pattern is to increment argc/argv after the getopt()
loop, regardless of whether or not the values are subsequently used.
Quit doing that, and remove the unused variables, it makes Coverity
sad.

Second, __wt_thread_join() was taking a copy of the __wt_thread_t
structure instead of a reference to the actual structure, which
means updates to the copy were never used. Switch __wt_thread_join
to take a reference to the actual structure so we're updating it
appropriately.

Change __wt_thread_join() to clear wt_thread_t.created in all error
paths rather than only on success. There's no point to repeatedly
trying to join a thread after failure.

Leave the unused value in __wt_txn_recover(), but use the standard
WiredTiger macro to clarify the issue.

test/bloom/test_bloom.c wasn't checking for illegal arguments set on
the command, add the usual test.