WT-4186 Log recovery/salvage now detects/reports corruption within log records.
This is done via a subtle trick: if there are two or more bytes of padding
at the end of the log record, set the final byte of padding to a non-zero
value. On the input side, we already know that a non-zero last byte means
that the write of the log record is a complete write. Thus, any checksum
mismatch for records with this property can be confidently reported as
data corruption.

Test case has been updated accordingly.