WT-4195 when encountering an illegal value, log the value that failed (#4184)
__wt_illegal_value() is used in a few places to display an additional
message, and most of them aren't a good use of that function.  Remove
the additional message string from __wt_illegal_value(), and replace
it in places where it makes sense with a varian of the WT_PANIC macro.

It's a program error to ever call __wt_btcur_insert_check() on anything
other than a row-store tree, assert it never happens rather than handling
the error.

Consolidate the error message and the illegal-value panic call into a
single failure for simplicity.