WT-3968 Use compression ratio to tune page sizes
Don't assume leaf page sizes are larger than internal page sizes when
calculating the default in-memory page image size.

Error out if a configured in-memory page image size is smaller than an
on-disk page size.

Don't do compression adjustment if the page image size is the same size
as the on-disk page size.