WT-4215 Allow recovery of log files from a backup without salvage. (#4205)
Because they are being read concurrently with being written, log files copied as part of a backup may have various portions that are zeroed. In particular, some log records may have their first part zeroed, and subsequent parts filled in. This should never happen for log files used after an application crash, so we must apply different rules for these cases. For backup log files, we allow checksums to fail, or holes in log files to end without the start of a valid record. Either one of these 'corruptions' is unreported, but causes the log to truncate at that point.