WT-4104 Records reappear after truncate when reads race with a checkpoint (#4210)
If an application does a search in a range that was deleted by a truncate while a tree is marked clean (in most applications, this is a short window during a checkpoint), then records can be resurrected.  Always mark pages dirty in that case, unless the tree is read-only (in which case the page reference should go back to deleted when it is evicted).