WT-4144 Fix rollback_to_stable with lookaside history. (#4212)
There were several bugs, resulting in incorrect data after a
rollback_to_stable call in the presence of lookaside history.

The main issue was with pages images containing updates newer than the
stable timestamp (which is a common case because WiredTiger evicts
optimistically, assuming the stable timestamp will move forward by the
time the next checkpoint starts).  Since rollback_to_stable doesn't
modify page images, it instead was marking such pages "invalid" so the
history would always be loaded before the pages could be read correctly.
However, the required history to make sense of such pages could be
removed by a lookaside sweep, which makes decisions solely on the
contents of the lookaside table without access to the "invalid" flag.

The solution here is to read the history into cache for such pages
during rollback_to_stable.  The page is marked dirty, with updates
newer than the stable timestamp rolled back, then it can be evicted or
rewritten in memory.  This eliminates the special "invalid" flag.

In the process of finding that bug, a number of other issues were found
and fixed, including:
* don't read all lookaside content into cache during rollback_to_stable;
* avoid races between eviction threads and rollback_to_stable (fix
  nesting of locks and transactions, use an explicit flag check during
  eviction rather than relying on the eviction server thread to update
  cache-wide flags); and
* simplify functions accessing the lookaside table so each function has
  a single session.