WT-4262 In eviction of internal pages, lock deleted children.
Elsewhere we assume that flipping ref->state from DELETED to LOCKED
makes it safe to free ref->page_del.  If we're attempting to evict an
internal page, we have it locked, so no thread can descend into it, but
a walk could be attempting to couple into a child page, so we can't
assume exclusive access to all of the children.