WT-4264 compaction can race with page modifications
Compaction can race with page modifications, that is, the WT_PAGE.modify
structure may be NULL and be created while compaction is testing it.
Don't set a local variable to the value of WT_PAGE.modify until we know
it exists.

Additionally, we shouldn't need to lock the page while compaction is looking
at it. We've blocked checkpoints from reconciling the page and we're holding
a hazard pointer on the WT_REF to prevent eviction from reconciling the page,
no other locks should be necessary.