WT-4267 fixed-length column store operations can corrupt data
A positioned cursor can be positioned in a set of implicitly created
records, which isn't correct (for example, WT_CURSOR.remove on an
implicitly created record should remove the implicitly created record,
not the location where the cursor is positioned).

It's easier to simply turn off pinned-page insert, update or remove
actions than to figure out if the cursor is correctly positioned, if
this turns out to be a performance problem, we can probably figure out
if we're correctly positioned by comparing the cursor's record number
to the pointed-at WT_UPDATE's record number.