WT-4177 Change log files when opening a backup cursor.
Applications that build their own log on top of WiredTiger may make writes that
are causally related to checkpoints it takes (such as log truncation). In those
cases, the writes should only be visible when restoring from a checkpoint that
triggered the write. Otherwise, WiredTiger may purge data required for
application level recovery.