WT-4385 Prepare-conflict during a cursor scan can return the wrong key (#4331)
* When returning a key after some number of prepare-conflicts, make sure
we don't use a pre-built key from a previous search. The underlying
key-return function uses a comparison of 0 to indicate the search function
has pre-built a key, set the cursor compare field to a non-zero value
to avoid that path.

* Add a check for cursor return ordering when eventually satisfying a cursor
scan after returning prepare-conflict.

* Use the cursor-return helper function instead of the explicit underlying
functions.

* Don't drop into the iteration code if we ever see an invalid value
when an update becomes visible after prepare-conflict returns during
a scan, that will result in an out-or-order key return, by definition.

The update value should be valid, assert that fact instead.