WT-4510 Failure when verifying internal pages of older version data files
When unpacking a WT_CELL_VALUE_COPY cell, unpack.cell points to the original cell, not the copied cell, because callers calculate offsets based on the cell location when setting up in-memory references to the page. This is a nasty little change in how cell unpack worked in
previous versions, required because the caller of the unpack code is no longer the "owner" of the current "cell" location in memory.