Abstract out address timestamp triplet initialization and add a comment.
Minor restructuring so the address timestamp triplet update function is
next to, and named similarly, to the initialization function.

Change the newest-start initialization to be WT_TS_NONE in all cases,
that's correct when maintaining timestamps (it will get updated as we
review key/value pairs), and correct when not maintaining timestamps
(all key/value pairs on the page are durable).