We have to use PRId64 (signed 8B integer), formats when writing the metadata records, because that's what the configuration code can handle. Configuration uses strtoll() to parse these strings, which means if we write UINT64_MAX as a PRIu64, we might get a saturated value (2^63), back.  (And of course, who knows what other strtoll() implementations do, it seems to me that an ERANGE failure isn't out of the question.)
This is a bug in existing metadata files in the field, that is, if any
of the fields we write are larger than 2^63, we'll parse them incorrectly.

Fortunately, the fields of interest are: the time (ckpt->sec), the size
(ckpt->ckpt_size), and the write-generation (ckpt->write_gen), none of
which are going to exceed 2^63 in any existing install.