Switch the cell descriptor byte's 0x08 bit to mean "timestamps set", that is, instead of the default being timestamped values, the default is a durable value and the bit indicates there are associated timestamps.
This simplifies things: first, that bit was 0 in historic pages, so
there's no need to test the page-version and the bit, we can just test
the bit.

Second, if a bit indicates a value is durable, then short values migrate
to 5-bits of encoded length instead of 6, which was going to be messy.
Instead, with this change, values with timestamps cannot be short values,
but over time, values that become globally visibile will migrate to
being short values.