WT-4595 Coverity "null pointer dereference" complaints (#4489)
* Set side-effect returns first in the function, before local variable
initialization.

Whitespace.

* WT_TRET() doesn't make sense after __wt_set_return(), that function
always returns an error. Make the __wt_session_release_dhandle()
function fail primary because it's a real failure we need to see,
and return EINVAL if the handle release succeeds.

* Replace WT_TRET with WT_ERR, otherwise we'll lose any error from
__wt_schema_release_table().

* Replace WT_TRET with WT_ERR, otherwise we'll lose any error from
__wt_schema_release_table().

* Rework __wt_schema_release_table() to return immediately if there's no
work to be done, and to set the table reference to NULL to ensure the
reference is never used after being released.