WT-4324 Ensure checkpoints rewrite pages with data in the future (#4521)
Sometimes we attempt eviction of a page by writing the most recent
versions of updates.  If such a page remains dirty in cache, it must be
visited by the next checkpoint to make sure the data that appears in the
checkpoint is a consistent cut of versions.

(cherry picked from commit d344940895bbf0232ea02e1e56fefac65d1eb2ae)