WT-4644 column-store missing run-length encoding opportunities (#4529)
Reconciliation can't reset the start/stop timestamps based on global visibility when laying out cells on the page because variable-length column store has already checked if the item matches previous items, that is, the RLE for the item is set, and if we change the timestamps, we might create sequential items that should have been RLE compressed.  Instead, the last time at which we can check for global visibility is when retrieving the values from the WT_UPDATE structure. This reverts
half of c04731f.

When stepping through a variable-length column store page, reset the start/stop timestamps to the on-page source values on every iteration, otherwise a single WT_UPDATE structure will cause the timestamps for future copies of the original on-page value to be incorrect.