WT-4695 Python3: allow most tests to run with Python3 with small changes (#4548)
Includes changes test suite and python API to allow most tests to run w/ Python3 (with a small additional change).  The test suite still works as is on Python2. Python3 requires callers to use bytes type when the key or value format is "u", this accounts for the bulk of the needed changes.