WT-4352 Resolve birthmarks during eviction in more cases.
Previously, we would resolve birthmarks for lookaside eviction, and in
some other cases (e.g., when overwriting an overflow value).  In
particular, we didn't resolve birthmarks during "scrub" eviction, where
a new disk image is calculated but kept in cache along with active
updates.

Since birthmarks assign version information to on-page data, they should
be resolved as part of any eviction that saves updates and overwrites
the previous on-page value.