Salvage checkpoints need to set WT_CKPT.metadata for the block manager, it's required and we'll drop core without it.
Salvage checkpoints were not being correctly initialized. I don't think
it was a problem in the field, but it wasn't correct, either. At the
least, the order and seconds should be set, and timestamps and transaction
values should be initialized.

There's no reason to build the salvage checkpoint separately from the
salvage code, merge them.