Add WT_BTREE_MIN_ALLOC_SIZE as a #define for 512B, the smallest WiredTiger allocation size.
Don't skip the descriptor block when scanning for live pages. The read will
fail, of course, but there's no gain in skipping the first 512B of the file.

Revert a change in __wt_open() to check for an absolute path name. It's not
needed, a simliar check appears in the underyling __wt_nfilename() function.

Review comments, mostly textual/comment cleanups, some whitespace.