If previous checkpoints are being deleted as part of a checkpoint, the file metadata from the previous checkpoint can't be repaired, we have to have the file metadata that is about to be stored in the database metadata file in the file when the current checkpoint completes, otherwise it won't be correct.
is a reasonably large change: previously we were storing the base file
metadata and the corrected checkpoint information in the file. Now,
the block manager now has to reach up into the metadata layer, construct
the complete almost-correct file checkpoint metadata entry, and copy it
into the file's avail list along with the base metadata and the corrected
checkpoint information.