WT-4878 Disable random dhandle selection for eviction (#4716)
This mostly goes back to the state before data handles were chosen randomly to search for eviction candidates.  It also changes some of the calculations for how many pages to enqueue per tree.

Also fix handle lock test to reset cursors before caching them. Otherwise, for configurations with many threads it can end up in a state where there is a cursor on every page, preventing the pages from splitting, and ultimately blocking eviction.  This was exposed by changes to eviction, but is a structural problem with the test case.