Increase the size of the repeatable-read buffer to from 512 to 256 slots.
Don't track reads made on a checkpoint handle for repetition.

Don't start a transaction if we're reading from a checkpoint. It's not
an error, but it's not an interesting test case.

Use tinfo->snap_first == tinfo->snap to flag if there are operations
to be updated at transaction resolution. It's not perfect, but the
failure mode is ignoring a set of operations which is fine.