WT-4733 bug fixes to long-term repeatable read branch (#4697)
* The repeatable read logic is wrong and we aren't repeating historic reads
as intended. There's no second out-of-band value for a timestamp and
it's confusing to try and fake it: add a boolean to the TINFO structure
that flags if an operation is repeatable.

* WiredTiger logs a verbose message when the read timestamp is set to a
value older than the oldest timestamp. Ignore the message, it happens
when repeating operations to confirm timestamped values don't change
underneath us.

* Yield whenever we can't start a transaction due to WT_CACHE_FULL.

If we find one of our timestamps is too old,  clear any operations with
timestamps at or older than the failing timestamp, that should help us
avoid wasting time on reads that won't work.

* WT-4694 Throttle reads in wtperf to make readonly workload perform more consistently (#4692)

* The final write in a checkpoint without an avail list will still write (#4696)

an empty avail list because the checkpoint needs to write the file
metadata. Fix the wt2336_base/base.run file to expect an additional write.

* Fix a comment, clarify a comparison.

* Split out the snap code into its own file.

* Refactor the snap repeatable read code.

* Split out the snap code into its own file.

* Temporarily stop repeating reads when reading from a checkpoint handle.

* WT-4863 Prototype WiredTiger inline functions (#4702)

* WT-4798 Fix bug in picking random dhandle for eviction (#4703)

* Add a #define instead of hard-coding the size of the WT_REF.hist array.
Shrink the size of the array elements, the line number and state are
relatively small.

* Instead of always choosing a current timestamp (that is, a timestamp
taken from the global timestamp allocator, which might be ahead of the
timestamps chosen when other transactions commit), choose a timestamp
that's known to be older than any updating transaction half the time.
That allows reads from the transaction to be safely repeated after the
transaction has completed. To support this, move the read and commit
timestamps from the ops() function stack into the TINFO structure, we
need the commit timestamp for later review.

Split out the code that creates a new WT_SESSION/WT_CURSOR pair into
a separate function.  Move the worker thread WT_SESSION/WT_CURSOR pair
into the TINFO structure.

* include/txn.i:42:43: error: implicit conversion loses integer precision:
'int' to 'uint16_t' (aka 'unsigned short') [-Werror,-Wconversion]

* ops.c:455:22: error: variable 'ts' may be uninitialized when used here
[-Werror,-Wconditional-uninitialized]

* Drop back to the original save-max value, I've been using a bigger number
for testing, and it got pushed.

* WT-4869 When eviction isn't keeping up, stop adding pages to cache. (#4699)

* WT-4877 Uninitialized memory being written during checkpoints (#4712)

When setting up a checkpoint, we leave room in the buffer for a packed uint64_t to be inserted later, and the packed uint64_t doesn't necessarily require all the space we allocate for it, and the remaining bytes were left uninitialized.

* WT-4875 fix commit timestamp assert to consider non transactional (#4706)

* WT-4875 fix commit timestamp assert to consider non transactional
tombstones.

* removed extra white space

* WT-4821 Update evergreen config to pull correct source for largescale test (#4717)

* WT-4878 Disable random dhandle selection for eviction (#4716)

This mostly goes back to the state before data handles were chosen randomly to search for eviction candidates.  It also changes some of the calculations for how many pages to enqueue per tree.

Also fix handle lock test to reset cursors before caching them. Otherwise, for configurations with many threads it can end up in a state where there is a cursor on every page, preventing the pages from splitting, and ultimately blocking eviction.  This was exposed by changes to eviction, but is a structural problem with the test case.

* Clean up format logging a bit -- log the values being read, remove the -L
option (output has long been too voluminous to not log it to a file), turn
off logging when we hit an error, reduce output in the next/prev log messages.

* Regularize log records to always include a row number, it makes it easier
to figure them out.

* Increase the size of the repeatable-read buffer to from 512 to 256 slots.

Don't track reads made on a checkpoint handle for repetition.

Don't start a transaction if we're reading from a checkpoint. It's not
an error, but it's not an interesting test case.

Use tinfo->snap_first == tinfo->snap to flag if there are operations
to be updated at transaction resolution. It's not perfect, but the
failure mode is ignoring a set of operations which is fine.

* typo

* Note that setting a timestamp for a transaction including a checkpoint
cursor is allowed but has no effect.

* Add logging for transaction begin/resolution, including timestamp
information.